+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+## 2.x
+
+* Added History extension
+* Added `ready` event
+* Added `loaded` event
+* Added `rendered` event
+* Added priority to callbacks
+* Added `initialize` call for extensions
+
 ## 2.0.0
 
 * Completely rewritten