+++ CHANGELOG.md
@@ -5,10 +5,11 @@ Changelog
 
 * Added History extension
 * Added `ready` event
-* Added `loaded` event
-* Added `rendered` event
+* Added `loaded` event (`load` is now triggered before loading starts)
+* Added `rendered` event (`render` is now triggered before rendering starts)
 * Added priority to callbacks
 * Added `initialize` call for extensions
+* Added `one` method
 
 ## 2.0.0